+++ src/main/java/com/learnway/study/service/StudyPostService.java
@@ -44,7 +44,7 @@ public class StudyPostService {
 	public Study boardadd(StudyDto dto) {
 		
 		//로그인한 값 가져와 수정예정
-		Member member = memberRepository.findById(1)
+		Member member = memberRepository.findById((long) 1)
 	            .orElseThrow(() -> new IllegalArgumentException("Invalid member ID: " + "1"));
 		
 		//멤버엔티티 연결해야함
